+++ web/apps/accounts/src/pages/_app.tsx
@@ -3,27 +3,21 @@ import { CustomHead } from "@/base/components/Head";
 import { LoadingOverlay } from "@/base/components/LoadingOverlay";
 import { AttributedMiniDialog } from "@/base/components/MiniDialog";
 import { useAttributedMiniDialog } from "@/base/components/utils/dialog";
-import { useSetupI18n } from "@/base/components/utils/hooks-i18n";
+import { useSetupI18n, useSetupLogs } from "@/base/components/utils/hooks-app";
 import { getTheme, THEME_COLOR } from "@/base/components/utils/theme";
-import { disableDiskLogs } from "@/base/log";
-import { logUnhandledErrorsAndRejections } from "@/base/log-web";
 import "@fontsource-variable/inter";
 import { CssBaseline } from "@mui/material";
 import { ThemeProvider } from "@mui/material/styles";
 import { t } from "i18next";
 import type { AppProps } from "next/app";
-import React, { useEffect, useMemo } from "react";
+import React, { useMemo } from "react";
 import { AppContext } from "../types/context";
 
 const App: React.FC<AppProps> = ({ Component, pageProps }) => {
+    useSetupLogs({ disableDiskLogs: true });
+
     const isI18nReady = useSetupI18n();
     const { showMiniDialog, miniDialogProps } = useAttributedMiniDialog();
-
-    useEffect(() => {
-        disableDiskLogs();
-        logUnhandledErrorsAndRejections(true);
-        return () => logUnhandledErrorsAndRejections(false);
-    }, []);
 
     const appContext = useMemo(() => ({ showMiniDialog }), [showMiniDialog]);
