+++ web/apps/cast/src/pages/_app.tsx
@@ -1,19 +1,14 @@
 import { staticAppTitle } from "@/base/app";
 import { CustomHead } from "@/base/components/Head";
+import { useSetupLogs } from "@/base/components/utils/hooks-app";
 import { getTheme, THEME_COLOR } from "@/base/components/utils/theme";
-import { disableDiskLogs } from "@/base/log";
-import { logUnhandledErrorsAndRejections } from "@/base/log-web";
 import "@fontsource-variable/inter";
 import { CssBaseline, ThemeProvider } from "@mui/material";
 import type { AppProps } from "next/app";
-import React, { useEffect } from "react";
+import React from "react";
 
 const App: React.FC<AppProps> = ({ Component, pageProps }) => {
-    useEffect(() => {
-        disableDiskLogs();
-        logUnhandledErrorsAndRejections(true);
-        return () => logUnhandledErrorsAndRejections(false);
-    }, []);
+    useSetupLogs({ disableDiskLogs: true });
 
     return (
         <>
